+++ include/ambition/7kaaInterface/input.hh
@@ -26,6 +26,8 @@
 #pragma once
 
 class GetA;
+class UnitCaravan;
+class VBrowseIF;
 
 
 namespace _7kaaAmbitionInterface::Input {
@@ -39,6 +41,14 @@ unsigned int cancelKeyEvent(
 );
 
 unsigned int confirmKeyEvent(
+);
+
+void detectCaravanCloneButton(
+  const UnitCaravan* _7kaaCaravan
+);
+
+bool detectTradeReportCaravanCloneButton(
+  VBrowseIF& caravanBrowser
 );
 
 bool detectClipboardKeys(