+++ CHANGELOG.md
@@ -12,6 +12,11 @@ speaking, considered to be the player's experience.
 
 ### Added
 
+- Added a clone button to Caravans.  If there is an idle Caravan, it is used as
+  the clone.  Otherwise, a new Caravan spawns at the Market closest to the
+  Caravan's first stop.
+- Added a caravan clone button to the trade report, next to the paste route
+  button.
 - Show the skill icon of the unit being trained in Towns under the progress bar
   on the world map and in the info pane.
 
@@ -19,6 +24,7 @@ speaking, considered to be the player's experience.
 
 - Fixed the Town training icon in the info pane not appearing until clicking on
   the Town again and disappearing after one unit finishes training.
+- Fixed graphics corruption with disabled buttons.
 
 
 ## [3.1.7] — 2025-06-13